Update Query pencarian
